Fixed Blat Search endpoint to use referenceId instead of BAM file id. Moved to POST /externaldb/blat/search endpoint
