+++ server/catgenome/src/main/java/com/epam/catgenome/controller/bam/BamController.java
@@ -30,11 +30,8 @@ import static com.epam.catgenome.controller.vo.Query2TrackConverter.convertToTra
 import java.io.IOException;
 import java.util.List;
 
-import com.epam.catgenome.controller.vo.ReadSequenceVO;
 import com.epam.catgenome.entity.bam.BamFile;
-import com.epam.catgenome.entity.bam.PSLRecord;
 import com.epam.catgenome.entity.bam.Read;
-import com.epam.catgenome.exception.ExternalDbUnavailableException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpStatus;
@@ -226,20 +223,5 @@ public class BamController extends AbstractRESTController {
                                  @RequestParam(required = false) final String fileUrl,
                                  @RequestParam(required = false) final String indexUrl) throws IOException {
         return Result.success(bamManager.loadRead(query, fileUrl, indexUrl));
-    }
-
-    @ResponseBody
-    @RequestMapping(value = "/bam/read/blat", method = RequestMethod.POST)
-    @ApiOperation(
-            value = "Returns statistics generated by BLAT search.",
-            notes = "Provides statistics generated by BLAT search performed on a read sequence.",
-            produces = MediaType.APPLICATION_JSON_VALUE)
-    @ApiResponses(
-            value = {@ApiResponse(code = HTTP_STATUS_OK, message = API_STATUS_DESCRIPTION)
-            })
-    public Result<List<PSLRecord>> blatReadSequence(@RequestParam final Long bamTrackId,
-                                                    @RequestBody final ReadSequenceVO readSequence)
-            throws IOException, ExternalDbUnavailableException {
-        return Result.success(bamManager.findBlatReadSequence(bamTrackId, readSequence.getReadSequence()));
     }
 }
