+++ server/catgenome/src/main/java/com/epam/catgenome/controller/externaldb/ExternalDBController.java
@@ -26,9 +26,13 @@ package com.epam.catgenome.controller.externaldb;
 
 import static com.epam.catgenome.component.MessageHelper.getMessage;
 
+import java.io.IOException;
 import java.util.List;
 import java.util.stream.Collectors;
 
+import com.epam.catgenome.controller.vo.ReadSequenceVO;
+import com.epam.catgenome.entity.bam.PSLRecord;
+import com.epam.catgenome.manager.bam.BlatSearchManager;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.MediaType;
 import org.springframework.stereotype.Controller;
@@ -37,6 +41,7 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
+import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.ResponseBody;
 
 import com.epam.catgenome.constant.MessagesConstants;
@@ -109,6 +114,9 @@ public class ExternalDBController extends AbstractRESTController {
 
     @Autowired
     private NCBIAuxiliaryManager ncbiAuxiliaryManager;
+
+    @Autowired
+    private BlatSearchManager blatSearchManager;
 
     @ResponseBody
     @RequestMapping(value = "/externaldb/uniprot/{geneId}/get", method = RequestMethod.GET)
@@ -355,5 +363,22 @@ public class ExternalDBController extends AbstractRESTController {
 
         List<Variation> variations = ncbiShortVarManager.fetchVariationsOnRegion(species, start, finish, chromosome);
         return Result.success(variations, getMessage(SUCCESS, NCBI));
+    }
+
+    @ResponseBody
+    @RequestMapping(value = "/externaldb/blat/search", method = RequestMethod.POST)
+    @ApiOperation(
+        value = "Returns statistics generated by BLAT search.",
+        notes = "Provides statistics generated by BLAT search performed on a read sequence. ReferenceId is required "
+                + "to search for required species",
+        produces = MediaType.APPLICATION_JSON_VALUE)
+    @ApiResponses(
+        value = {@ApiResponse(code = HTTP_STATUS_OK, message = API_STATUS_DESCRIPTION)
+        })
+    public Result<List<PSLRecord>> blatReadSequence(@RequestParam
+                                                    final Long referenceId,
+                                                    @RequestBody final ReadSequenceVO readSequence)
+        throws IOException, ExternalDbUnavailableException {
+        return Result.success(blatSearchManager.findBlatReadSequence(referenceId, readSequence.getReadSequence()));
     }
 }
